requirements.txt
Pillow
coverage